+++ Source/Bifrost.Web.Mvc.Specs/Validation/for_ValidatorPropertyValidator/when_creating_with_validator_with_dynamic_state.cs
@@ -1,6 +1,6 @@
-﻿using Bifrost.Web.Mvc.Validation;
+﻿using System.Web.Mvc;
+using Bifrost.Web.Mvc.Validation;
 using Machine.Specifications;
-using System.Web.Mvc;
 using Moq;
 using It = Machine.Specifications.It;