+++ Source/Bifrost.Web/Configuration/ConfigurationRouteHttpHandler.cs
@@ -103,10 +103,10 @@ namespace Bifrost.Web.Configuration
             var builder = new StringBuilder();
 
             if (_webConfiguration.ScriptsToInclude.JQuery)
-                builder.Append(GetResource("Bifrost.Web.Scripts.jquery-2.1.3.min.js"));
+                builder.Append(GetResource("Bifrost.Web.Scripts.jquery-2.2.3.min.js"));
 
             if (_webConfiguration.ScriptsToInclude.Knockout)
-                builder.Append(GetResource("Bifrost.Web.Scripts.knockout-3.2.0.debug.js"));
+                builder.Append(GetResource("Bifrost.Web.Scripts.knockout-3.4.0.debug.js"));
 
             if (_webConfiguration.ScriptsToInclude.SignalR)
                 builder.Append(GetResource("Bifrost.Web.Scripts.jquery.signalR-2.2.0.js"));
@@ -117,9 +117,7 @@ namespace Bifrost.Web.Configuration
             if (_webConfiguration.ScriptsToInclude.Require)
             {
                 builder.Append(GetResource("Bifrost.Web.Scripts.require.js"));
-                builder.Append(GetResource("Bifrost.Web.Scripts.order.js"));
                 builder.Append(GetResource("Bifrost.Web.Scripts.text.js"));
-                builder.Append(GetResource("Bifrost.Web.Scripts.domReady.js"));
                 builder.Append(GetResource("Bifrost.Web.Scripts.noext.js"));
             }