+++ Source/Bifrost.Web/SignalR/BifrostDependencyResolver.cs
@@ -41,29 +41,14 @@ namespace Bifrost.Web.SignalR
                 catch { }
 
             return service;
-
-            /*
-            if (!IsSignalRInternalType(serviceType))
-            //if( _container.HasBindingFor(serviceType) )
-                return _container.Get(serviceType);
-
-            return base.GetService(serviceType);*/
         }
 
         public override IEnumerable<object> GetServices(Type serviceType)
         {
-            //if (!IsSignalRInternalType(serviceType) )
             if (_container.HasBindingFor(serviceType))
                 return _container.GetAll(serviceType).Concat(base.GetServices(serviceType));
 
             return base.GetServices(serviceType);
         }
-
-
-        bool IsSignalRInternalType(Type serviceType)
-        {
-            return serviceType.Namespace.StartsWith("Microsoft") || serviceType.Namespace.StartsWith("System") || serviceType.Namespace.StartsWith("Owin");
-        }
-
     }
 }